Conflicts:
	MonsterSlide/src/fh/teamproject/entities/World.java

+++ MonsterSlide/src/fh/teamproject/utils/Settings.java
@@ -4,7 +4,7 @@ public class Settings {
 
 	public float SLIDE_WIDTH;
 	public float SLIDE_LENGTH;
-	public float SLIDE_SEGMENTS;
+	public int SLIDE_SEGMENTS;
 	public float SLIDE_SLOPE;
 	public float SLIDE_DIFFICULTY;
 	public float SLIDE_DIMENSION;
